fjerner ubrukte komponenter og metoder fra journalføring-1.0
