+++ src/frontend/App/context/toggles.ts
@@ -18,6 +18,5 @@ export enum ToggleName {
     // Midlertidige toggles - kan fjernes etterhvert
     visSatsendring = 'familie.ef.sak.frontend-vis-satsendring',
     visInntektPersonoversikt = 'familie.ef.sak.frontend.vis-inntekt-personoversikt',
-    visNyJournalføring = 'familie.ef.sak-ny-journalforing',
     visKAVedtakAlternativ = 'familie.ef.sak.frontend.vis-ka-uten-brev',
 }
