+++ src/frontend/Komponenter/Journalføring/Felles/utils.ts
@@ -1,17 +1,8 @@
-import {
-    DokumentTitler,
-    IJournalpostResponse,
-    LogiskVedlegg,
-} from '../../../App/typer/journalføring';
-import {
-    Behandlingstema,
-    behandlingstemaTilTekst,
-    Stønadstype,
-} from '../../../App/typer/behandlingstema';
+import { IJournalpostResponse, LogiskVedlegg } from '../../../App/typer/journalføring';
+import { Stønadstype } from '../../../App/typer/behandlingstema';
 import { Behandling, BehandlingResultat } from '../../../App/typer/fagsak';
 import { Behandlingstype } from '../../../App/typer/behandlingstype';
-import { BehandlingRequest, Journalføringsaksjon } from '../../../App/hooks/useJournalføringState';
-import { BehandlingKlageRequest } from '../../../App/hooks/useJournalføringKlageState';
+import { Journalføringsaksjon } from '../../../App/hooks/useJournalføringState';
 import { ISelectOption, MultiValue, SingleValue } from '@navikt/familie-form-elements';
 import { Klagebehandlinger } from '../../../App/typer/klage';
 
@@ -19,34 +10,8 @@ export const JOURNALPOST_QUERY_STRING = 'journalpostId';
 export const OPPGAVEID_QUERY_STRING = 'oppgaveId';
 export type MultiSelectValue = { label: string; value: string };
 
-export const lagJournalføringKlageUrl = (
-    journalpostId: string,
-    oppgaveId: string | number
-): string => {
-    return `/journalfor-klage?${JOURNALPOST_QUERY_STRING}=${journalpostId}&${OPPGAVEID_QUERY_STRING}=${oppgaveId}`;
-};
-
-export const lagJournalføringUrl = (journalpostId: string, oppgaveId: string | number): string => {
-    return `/journalfor?${JOURNALPOST_QUERY_STRING}=${journalpostId}&${OPPGAVEID_QUERY_STRING}=${oppgaveId}`;
-};
-
-export const harTittelForAlleDokumenter = (
-    journalResponse: IJournalpostResponse,
-    dokumentTitler?: DokumentTitler
-) =>
-    journalResponse.journalpost.dokumenter
-        .map((d) => d.tittel || (dokumentTitler && dokumentTitler[d.dokumentInfoId]))
-        .every((tittel) => tittel && tittel.trim());
-
-export const utledKolonneTittel = (
-    behandlingsTema: Behandlingstema | undefined,
-    fra: 'klage' | 'vanlig'
-) => {
-    const prefix = 'Registrere journalpost';
-    const type = fra === 'klage' ? ' for klage' : '';
-    const stønadstype = behandlingsTema ? ': ' + behandlingstemaTilTekst[behandlingsTema] : '';
-    return `${prefix}${type}${stønadstype}`;
-};
+export const lagJournalføringUrl = (journalpostId: string, oppgaveId: string | number): string =>
+    `/journalfor?${JOURNALPOST_QUERY_STRING}=${journalpostId}&${OPPGAVEID_QUERY_STRING}=${oppgaveId}`;
 
 export const utledRiktigBehandlingstype = (
     tidligereBehandlinger: Behandling[]
@@ -59,13 +24,6 @@ export const utledRiktigBehandlingstype = (
         ? Behandlingstype.REVURDERING
         : Behandlingstype.FØRSTEGANGSBEHANDLING;
 };
-
-export const harValgtNyBehandling = (behandling: BehandlingRequest | undefined): boolean =>
-    behandling !== undefined && behandling.behandlingsId === undefined;
-
-export const harValgtNyKlageBehandling = (
-    behandling: BehandlingKlageRequest | undefined
-): boolean => behandling !== undefined && behandling.behandlingId === undefined;
 
 const dokumentTitler: string[] = [
     'Anke på tilbakekreving',
@@ -171,3 +129,9 @@ export const skalViseBekreftelsesmodal = (
     journalføringsaksjon === Journalføringsaksjon.OPPRETT_BEHANDLING
         ? false
         : journalResponse.harStrukturertSøknad || erPapirSøknad;
+
+export enum UstrukturertDokumentasjonType {
+    PAPIRSØKNAD = 'PAPIRSØKNAD',
+    ETTERSENDING = 'ETTERSENDING',
+    IKKE_VALGT = 'IKKE_VALGT',
+}
